# Optimize: jwt regexp predicates