optimize: JWT regexp based predicates

Context:
Normally clients reuse their tokens so we can easily cache the matching result by token for each predicate.
This will speedup significantly FabricGateway matching lookups.

Adds a sync.Map that we use to perform a cache lookup of the wire string to the result.
A new type `registry` is used to cleanup all entries every hour. For simplicity we do not track timestamps. Tested by a benchmark to cleanup every Millisecond and the result does not change significantly, so we should not care about the recreation of the cache once in a while.

Benchmarks for benchstat comparision ran with count 20
```
goos: darwin
goarch: arm64
pkg: github.com/zalando/skipper/predicates/auth
cpu: Apple M2
                        │    old.txt    │               new.txt               │
                        │    sec/op     │   sec/op     vs base                │
JWTPayloadAllKVRegexp     4079.00n ± 1%   45.59n ± 0%  -98.88% (p=0.000 n=20)
JWTPayloadAllKVRegexp-2   3600.00n ± 0%   45.57n ± 0%  -98.73% (p=0.000 n=20)
JWTPayloadAllKVRegexp-4   3607.00n ± 0%   45.59n ± 0%  -98.74% (p=0.000 n=20)
JWTPayloadAllKVRegexp-8   3603.00n ± 0%   45.70n ± 0%  -98.73% (p=0.000 n=20)
geomean                     3.717µ        45.61n       -98.77%

                        │   old.txt    │                  new.txt                  │
                        │     B/op     │     B/op      vs base                     │
JWTPayloadAllKVRegexp     2.250Ki ± 0%   0.000Ki ± 0%  -100.00% (p=0.000 n=20)
JWTPayloadAllKVRegexp-2   2.250Ki ± 0%   0.000Ki ± 0%  -100.00% (p=0.000 n=20)
JWTPayloadAllKVRegexp-4   2.250Ki ± 0%   0.000Ki ± 0%  -100.00% (p=0.000 n=20)
JWTPayloadAllKVRegexp-8   2.250Ki ± 0%   0.000Ki ± 0%  -100.00% (p=0.000 n=20)
geomean                   2.250Ki                      ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                        │  old.txt   │                new.txt                 │
                        │ allocs/op  │ allocs/op  vs base                     │
JWTPayloadAllKVRegexp     51.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=20)
JWTPayloadAllKVRegexp-2   51.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=20)
JWTPayloadAllKVRegexp-4   51.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=20)
JWTPayloadAllKVRegexp-8   51.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=20)
geomean                   51.00                   ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean
```


old:
```
% go test -run '^$' -benchmem -cpu 1,2,4,8 -bench BenchmarkJWTPayloadAnyKVRegexp ./predicates/auth -count 1
goos: darwin
goarch: arm64
pkg: github.com/zalando/skipper/predicates/auth
cpu: Apple M2
BenchmarkJWTPayloadAnyKVRegexp            311738              3985 ns/op            2304 B/op         51 allocs/op
BenchmarkJWTPayloadAnyKVRegexp-2          338103              3524 ns/op            2304 B/op         51 allocs/op
BenchmarkJWTPayloadAnyKVRegexp-4          343358              3519 ns/op            2304 B/op         51 allocs/op
BenchmarkJWTPayloadAnyKVRegexp-8          342327              3543 ns/op            2304 B/op         51 allocs/op
PASS
ok      github.com/zalando/skipper/predicates/auth      5.013s
```
new:
```
% go test -run '^$' -benchmem -cpu 1,2,4,8 -bench BenchmarkJWTPayloadAnyKVRegexp ./predicates/auth -count 1               
goos: darwin
goarch: arm64
pkg: github.com/zalando/skipper/predicates/auth
cpu: Apple M2
BenchmarkJWTPayloadAnyKVRegexp          24621844                45.77 ns/op            0 B/op          0 allocs/op
BenchmarkJWTPayloadAnyKVRegexp-2        26927236                45.68 ns/op            0 B/op          0 allocs/op
BenchmarkJWTPayloadAnyKVRegexp-4        26123988                45.59 ns/op            0 B/op          0 allocs/op
BenchmarkJWTPayloadAnyKVRegexp-8        26365248                45.65 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/zalando/skipper/predicates/auth      4.952s
```